requirements.txt
mutagen>=1.47.0
send2trash>=1.8.0
librosa>=0.10.0
numpy>=1.24.0
matplotlib>=3.7.0
pydub>=0.25.0
rapidfuzz>=3.0.0
python-Levenshtein>=0.21.0
